fix #7 - Criação de arquitetura base para elementos genéricos do board.
- Agora o board possui BoardElements, classe abstrata usada para representar qualquer coisa que ocupe um espaço no board.
- Tile e Obstacle são classes independentes que herdam de BoardElement.
- Vários métodos e mecanismos foram transformados em callbacks.
- Responsabilidades foram redistribuídas entre as classes.
