+++ Assets/Scripts/Board/Behaviour.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 419183e4cfda73f4690cbef01802d0fd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
