+++ Assets/Scripts/Board/Elements.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a6724d86ddf1e844cac710ed296c4597
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
