+++ Assets/Scripts/Board/Elements/Base.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f5376fe9ea6664f42b8676639b2a6b3a
+guid: aa7893a859dee6f4fa5c5d95c57bf1e1
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
