+++ Assets/Scripts/Board/Elements/Base/BoardElement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0ae0b74784d2ac54b9e82c7a3bf476fd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
