+++ Assets/Scripts/Board/Elements/Obstacle.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2533d44b1099f6d478bc49e99c24a4f9
+guid: 47b98e6e3bca2704c8bfa454a6fc5641
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
