+++ Assets/Scripts/Board/Elements/Tile.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 2868b16bba0b5904e8412c0512236014
-TextScriptImporter:
+guid: 85bfa924ce558c44ba6699d131363bcc
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
