+++ Assets/Scripts/Board/Elements/Tile/PowerTile.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: da228e0dbcf7fba4895e921184047a98
+guid: 3f95c444201048548aecdd0dac75f7e0
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
