+++ Assets/Scripts/Board/Elements/Tile/TileType.cs
@@ -1,0 +1,12 @@
+﻿namespace Board.Elements.Tile
+{
+    public enum TileType
+    {
+        Blue = 0,
+        Gray = 1,
+        Green = 2,
+        Red = 3,
+        Yellow = 4,
+        PowerUp = 5
+    }
+}
