+++ Assets/Scripts/Board/Elements/Tile/TileType.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 892f8d1708f349242bc76786e252c116
+guid: 299d01c2073418144bbe461f6dbec3a5
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
