--- Assets/Scripts/Enum/TileTypes.cs
@@ -1,10 +1,0 @@
-﻿public enum TileTypes
-{
-    Blue = 0,
-    Gray = 1,
-    Green = 2,
-    Red = 3,
-    Yellow = 4,
-    PowerUp = 5,
-    Obstacle = 6,
-}
