+++ Assets/Scripts/Power/Delegate.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3f892c3a6618e5a4da933133dd911979
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
